# backend/src/health/health-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class HealthResponseDto {
  @ApiProperty({ enum: ['ok'] })
  status!: 'ok';

  @ApiProperty({ enum: ['up'] })
  database!: 'up';

  @ApiProperty({ format: 'date-time' })
  checkedAt!: string;
}